Define databased models for the dataset
The dataset are from : https://fdc.nal.usda.gov/download-datasets.html

Latest release in October. The models create a simple version from the dataset and the following dataset file will be implemented : Food, FoodCategory, Nutrients, FoodNutrient.

This dataset are crucial for the application when the user will add new Meal in the website. The user will apply ingredients from the registered Food and Nutrient. The functionality after submitting all ingredients will properly calculate the result and be availabel in the website.

The FoodNutrient are Food associated with the Nutrient. Where several record from Food table has relationship with Nutrient table.