Create Two function in the dataset models
override the method equals -> to check the new object added to the database exist in database or not. This is crucial for database integrity to not store duplicated value.

Override GetHashMehtod in order to calculate the hash value of the properities, as it known the identical name has the same hash value, this are part of the integration to enusre the database integrity to not store duplicate value when creating sevral datasets