Create dataSetImporter
This file wills track the directory where the dataset are installed. It will apply the dataset to the models defined in the database. In this approach we apply the datasets file  to the models entitiy framqwork to further configurated with the application.

The map classes ->  track the column name of the table in the Entity framwork models to assoicated with the dataset column in order to make the import.