asd (#35)
* Update med kommentar LoginController.cs (#17)

Co-authored-by: arjan <97673925+AreArjane@users.noreply.github.com>

* update

* v

* f

* d

* Update med kommentar LoginController.cs (#21)

Co-authored-by: mahsaimani <114579479+mahsaimani@users.noreply.github.com>

* Update med kommentar LoginController.cs (#22)

Co-authored-by: mahsaimani <114579479+mahsaimani@users.noreply.github.com>

* Merge pull request #2 from AreArjane/branch-1 (#3)

* ;

* l (#23)

* l (#24)

* Update med kommentar AuthController.cs (#19)

Co-authored-by: arjan <97673925+AreArjane@users.noreply.github.com>

* Update HomeController.cs (#16)

* Update med kommentar SuperUserController.cs (#15)

* Update med kommentar MealController.cs (#14)

* Update med kommentar MealController.cs (#25)

Co-authored-by: mahsaimani <114579479+mahsaimani@users.noreply.github.com>

* Update med kommentar NormalUserController.cs (#26)

Co-authored-by: mahsaimani <114579479+mahsaimani@users.noreply.github.com>

* Update med kommentar LoginController.cs (#29)

Co-authored-by: mahsaimani <114579479+mahsaimani@users.noreply.github.com>

* Update med kommentar AuthController.cs (#30)

Co-authored-by: mahsaimani <114579479+mahsaimani@users.noreply.github.com>

* l (#31)

* Branch 2 (#32)

Co-authored-by: Shefaa62 <116075255+Shefaa62@users.noreply.github.com>

* sf (#33) (#34)

* Add Database Context

Using PostGres DataBases. Create DataBase Context file to configurated models with database migration.

* .

* Configurate Database with Neon Postgres

Create an enviroment variabel hold sensitiv information related to the database connection.

We add the enviroment variabel to the application.json file and configurated in the program.cs.

To transfer the database models to the Postgres Configuration.

builder.Services.AddDbContext<ApplicationDbContext>(options =>
    options.UseNpgsql(connectionString));
Console.WriteLine("Connection String: " + connectionString);

In order to run :

 dotnet-ef migrations add InitialCreate --context ApplicationDbContext

dotnet -ef database update

* .

* .

* .

* .

* .

* Create Two function in the dataset models

override the method equals -> to check the new object added to the database exist in database or not. This is crucial for database integrity to not store duplicated value.

Override GetHashMehtod in order to calculate the hash value of the properities, as it known the identical name has the same hash value, this are part of the integration to enusre the database integrity to not store duplicate value when creating sevral datasets

* Create dataSetImporter

This file wills track the directory where the dataset are installed. It will apply the dataset to the models defined in the database. In this approach we apply the datasets file  to the models entitiy framqwork to further configurated with the application.

The map classes ->  track the column name of the table in the Entity framwork models to assoicated with the dataset column in order to make the import.

* Try script method for

* update

* delete migrations

* Create Meal Models

In this models is it crucisal for the user to create the meal based on registered FOOD. We manipulate Nutient based on the MealFood Table .

For this approach : c

Meal  table create hold trace of the meal name
MealFood trace meal with the food included

* Update models

* Create Controller

Create Three Type of Controller :

Search Controller take the query from the user and filter agains Food and Meals.

FoodController -> manipulate to each single Food with the sets of the nutrients

MealContoller -> manipulate to each single meals with the sets of Food and nutients

* Update Readme

Add Controller description
Database script handling

* .

* Adding the UserType

Adding UserType to the abstracted class User with the enumration of each user in innheritance it will give more preformance to the function implementation such Filtering, and can make more strick and extra secure managment to Access Controll

* Manage directory structure

* Create the controller for the User

Create SuperUser and NormalUser Controller

* Modification to the UserController

* Delete damage file

* Update AdminCOntroller

Add admin retriving all user with pagination and filter Firstname letters

* Create Directory Structure for View Sub1

* .

* Update ReadMe

* Update View

* s

* change to Views

* Update views

* update

* Update Views

* Update user models with hash password

* f

* Update med kommentar FoodCategoryMap.cs

* Latest Update

* Update med kommentar LoginController.cs (#17)



* update

* v

* f

* d

* Update med kommentar LoginController.cs (#21)



* Update med kommentar LoginController.cs (#22)



* Merge pull request #2 from AreArjane/branch-1 (#3)

* ;

* l (#23)

* l (#24)

* Update med kommentar AuthController.cs (#19)



* Update HomeController.cs (#16)

* Update med kommentar SuperUserController.cs (#15)

* Update med kommentar MealController.cs (#14)

* Update med kommentar MealController.cs (#25)



* Update med kommentar NormalUserController.cs (#26)



* Update med kommentar LoginController.cs (#29)



* Update med kommentar AuthController.cs (#30)



* l (#31)

* Branch 2 (#32)



---------

Co-authored-by: mahsaimani <114579479+mahsaimani@users.noreply.github.com>
Co-authored-by: Shefaa62 <116075255+Shefaa62@users.noreply.github.com>

---------

Co-authored-by: mahsaimani <114579479+mahsaimani@users.noreply.github.com>
Co-authored-by: Shefaa62 <116075255+Shefaa62@users.noreply.github.com>